requirements.txt
requests==2.27.1
httpx==0.28.1
python-telegram-bot==22.2